完善上次的问题：点赞、收藏、分享、评论

+++ src/com/threeblog/dao/AnswersDao.java
@@ -42,7 +42,7 @@ public interface AnswersDao {
 	 * @return
 	 * @throws SQLException
 	 */
-	int findCommentAnswerNum(String comment_id)throws SQLException;
+	Long findCommentAnswerNum(String comment_id)throws SQLException;
 	
 	/**
 	 * 通过文章留言id删除留言下的回复
